test(get-and-delete-user): implement integration tests

+++ src/user/application/usecases/DeleteUser.usecase.ts
@@ -1,6 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { UseCase } from '../../../core/shared/helpers/usecase';
 import { UserRepository } from '../../domain/repositories/user.repository';
+import { validate } from 'uuid';
 
 @Injectable()
 export class DeleteUserUseCase implements UseCase {
@@ -10,6 +11,9 @@ export class DeleteUserUseCase implements UseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    if (!validate(id)) {
+      return;
+    }
     await this.userRepository.delete(id);
   }
 }
